requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
langgraph
langchain-core
langchain_groq
langchain_huggingface
langchain_text_splitters
langchain_community
faiss-cpu
pillow
